move `get_random_salt` to deployer.py
