+++ starknet_py/net/udc_deployer/deployer.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import random
 from typing import Optional, List, Union, cast, NamedTuple
 
 from starkware.starknet.public.abi import get_selector_from_name
@@ -9,12 +10,13 @@ from starknet_py.constants import DEFAULT_DEPLOYER_ADDRESS
 from starknet_py.net.client_models import Hash, Call
 from starknet_py.net.models import AddressRepresentation, parse_address, compute_address
 from starknet_py.utils.contructor_args_translator import translate_constructor_args
-from starknet_py.utils.crypto.facade import pedersen_hash, get_random_salt
+from starknet_py.utils.crypto.facade import pedersen_hash
 from starknet_py.utils.data_transformer.universal_deployer_serializer import (
     universal_deployer_serializer,
     deploy_contract_abi,
 )
 from starknet_py.utils.sync import add_sync_methods
+from starknet_py.constants import FIELD_PRIME
 
 
 ContractDeployment = NamedTuple("ContractDeployment", [("udc", Call), ("address", int)])
@@ -123,3 +125,7 @@ class Deployer:
             salt=salt,
             deployer_address=deployer_address,
         )
+
+
+def get_random_salt() -> int:
+    return random.Random().randrange(0, FIELD_PRIME)
