+++ starknet_py/utils/crypto/facade.py
@@ -1,5 +1,4 @@
 import os
-import random
 from typing import Optional
 
 from crypto_cpp_py.cpp_bindings import (
@@ -11,7 +10,6 @@ from starkware.cairo.common.hash_state import compute_hash_on_elements
 from starkware.cairo.lang.vm.crypto import pedersen_hash as default_hash
 from starkware.crypto.signature.signature import sign
 
-from starknet_py.constants import FIELD_PRIME
 from starknet_py.net.client_models import Call
 
 # PREFIX_TRANSACTION = encoded 'StarkNet Transaction'
@@ -48,7 +46,3 @@ def pedersen_hash(left: int, right: int) -> int:
     if use_cpp_variant():
         return cpp_hash(left, right)
     return default_hash(left, right)
-
-
-def get_random_salt() -> int:
-    return random.Random().randrange(0, FIELD_PRIME)
